following same pattern of simple LRU, separate 2q/arc Cache and LRU, make XXXCache just thread-safe wrapper over XXXLRU, avoid NoLock ugliness
